feat: add minimal logging capabilities
This commit adds a quick-and-dirty first cut of supporting
logging and logging configuration.  I chose to use the 'pino'
logging framework for now, because it is one of the more popular
JS frameworks, supports nodejs and browser, and seems to be
focused on performance.

We abstract the logging implementation behind our own facade so
that we can swap out pino later for a different library if
we decide we need to, without making any breaking changes to
our API.