fix: memory issues for retries; improvements to logging (#82)
* fix: memory issues for retries; improvements to logging

This commit fixes a memory issue that surfaced on the retry logic
under higher load.

It also adds a broader array of log levels for users to choose
from, and moves the noisiest log messages to `trace`.  In addition,
there are some places where we had been logging the cache value;
those are removed because for large cache values the log messages
were overwhelming.