chore: modify github action to do releases from 'release' branch (#83)
This commit changes the behavior so that we won't do a release to
npmjs on every commit to `main`.  We will be releasing from the
`release` branch from now on.  In a future commit we'll add
a manually-dispatched action to trigger a release, which will
involve doing a merge from `main` to `release` and then pushing
to `release` to trigger this action.