fix: simplify interceptors, pass cache name as metadata (#95)
This commit is modeled after the way the python code is set up.
Instead of passing the cache name as a constructor arg to the
HeaderInterceptor (which forces us to construct a new instance
of that interceptor for each request), we now specify a metadata
object directly on the request that we construct and we set
the cache name there.  This simplifies some of the code because
now we can construct the list of interceptors once at initialization.