chore: move max idle millis setting into transport strategy
This addresses one of the outstanding items from the list
of fast-follows for the Configuration API:

https://github.com/momentohq/client-sdk-javascript/issues/146

Prior to this commit we were exposing 'maxIdleMillis' as a
top-level setting on the Configuration.  However, we do not
expect users to need to know about this setting in the 90%
use case, so this commit moves it down into the TransportStrategy.