chore: move max idle millis setting into transport strategy (#176)
* chore: move max idle millis setting into transport strategy

This addresses one of the outstanding items from the list
of fast-follows for the Configuration API:

https://github.com/momentohq/client-sdk-javascript/issues/146

Prior to this commit we were exposing 'maxIdleMillis' as a
top-level setting on the Configuration.  However, we do not
expect users to need to know about this setting in the 90%
use case, so this commit moves it down into the TransportStrategy.

* chore: add Props interfaces and unit tests for config copy constructors (#177)

This commit makes a small tweak to the constructor signatures for
the various classes in the Configuration heirarchy; we now use
a Properties interface for each one so that arguments can be
passed in keyword-style.  This is with an eye towards some of these
argument lists expanding over time, and to make things a bit more
consistent / readable.

We also add unit test coverage for the copy constructors for all
of these classes, as was requested in an earlier PR.  This
addresses one of the fast-follow items from:

https://github.com/momentohq/client-sdk-javascript/issues/146

This change shouldn't have any API implications for existing users
as long as they were using one of the pre-built configs.