fix: Prevent infinite loop in example
Use loose equality when checking for the presence of a list caches
token so that null can be compared to undefined. Fixes an infinte
loop that is only broken by rate limiting.

Give the token to the list caches call to prevent an infinite loop
that would occur if there were more than one page of caches.