feat: Add RetryStrategy configuration
This commit introduces the RetryStrategy interface as part of
our Configuration object.  Also introduces the most basic
retry strategy, FixedCountRetryStrategy, and wires it up
to the RetryInterceptor.

This causes some quirkiness with logging initialization and
highlights that the logging API needs to be re-worked a bit.
Will tackle that in a follow-up PR; in the meantime the logger
will log at DEBUG level.