feat: re-work Logging API
Prior to this commit, our logging API was just a quick-and-dirty
first pass; it was very inflexible and tightly coupled to Pino.

In this commit we re-work the API to be more like the .NET one.
Customers provide a MomentoLoggerFactory that knows how to create
MomentoLogger objects, and we use those for our logging.  We
provide a default initial implementation based on Pino, but
we can trivially add new implementations that support any other
existing logging framework that a customer may be using, so that
they can be assured that Momento log messages will be routed
to the same destinations as any other log messages that they
are producing in their apps.