chore: swap truncate and ttl arguments
Swap the ttl and truncate arguments in the list methods. Ttl needs to be
specified less often than the truncate size, so it makes more sense to
have it at the end where it can be left off.

Change the name of the ttl argument to collectionTtl for all methods
that take a CollectionTtl object to avoid confusion with methods where
the ttl is a number.