fix: remove pino dependency for default logger (#241)
* fix: remove pino dependency for default logger

A customer ran into an issue with the pino dependency, and after
thinking about it I realized that we shouldn't be forcing that
transitive dependency on users.  This commit removes the pino
logger and updates the Default logger to just use basic
console.log statements.  We vendor an implementation of printf
to allow us to avoid interpolating strings when unnecessary.

In a subsequent commit I will add the Pino logger back in to
the examples directory, and in the future if users are interested
in specific pre-built logger configurations we can figure out how
to package them without requiring the transitive deps for all
users.