fix: remove withLoggerFactory from Configurations
I had forgotten a hard-fought lesson that I learned in the .NET
SDK (which also needs this fix): the pre-built configurations
will often pass the loggerFactory around to the constructors
of other components that they instantiate, such as the RetryInterceptor.

If this happens, and then we allow the user to override the loggerFactory
on the config with a new one, we can end up in a state where the original
factory is being used for some components and the new one for other
components.  For a short time I attempted to reconcile this by trying
to keep track of all of the places where the loggerFactory had been
referenced and then trying to update those as well, but it was a big
mess that wasn't guaranteed to work so we decided that was a bad path
to go down.

Hence, it is necessary for users to specify the one true loggerFactory
(or use the default one) at the time when they construct their
Configuration object.  To that end this commit removes the function
that would allow them to override it.