fix: remove shared exports of response types (#273)
* fix: remove shared exports of response types

This commit removes the 'response-simple-success' namespace and
fixes a few other places where we were doing `export * from '...'`,
and duplicates the relevant code into the response types that need
it.  This makes the code less DRY but it is necessary because
without this, editors get confused about the namespaces of the
types.  (e.g., IntelliJ would show the response type of `client.set`
as `CreateCache.Response` because that was an alias for the same
type.)

* fix: remove response-get-value as well, since it was only used once