WIP: sketch 'hitOrElse'
I don't think this actually works well because of the combinatorial
types due to byte array vs. string.  It's probably better left
as some code we could include in "advanced examples" or something,
and then let the reader implement it in a utils namespace based on
the types they care about.