chore: improve example code (#281)
This commit splits out the main example into two files: `basic.ts`
and `advanced.ts`.  The `basic.ts` is intended to be a quickstart
so we try to keep it as succint as possible.  The `advanced` example
adds more features such as listCaches, delete, and an example of how
to issue concurrent gets.