fix: bug in middleware reversing, plus add requestBodyTime to csv
This commit fixes a bug in the middleware order; previously I was
calling `.reverse()` every time, but I didn't realize that function
modified the array in place, which caused broken behavior.  This is
now fixed by doing a single clone and reverse at the beginning
of the function and then using the appropriate (original or reversed)
list for the appropriate life cycles.

Also adds requestBodyTime to the experimental metrics middleware.