chore: re-enable SortedSet tests that assert empty list response
Prior to this commit we had temporarily disabled some sorted
set tests that were asserting that an empty list would be
returned in cases where the set existed but the requested rank/offset
resulted in an empty list.  The server was returning a MISS
in those cases.  It has since been updated to the expected
behavior so now we re-enable the tests.