fix: wait for first response on pubsub and also check for cache deleted midstream (#358)
Updates the pubsub stream handling to wait for the first response before returning a TopicSubscribe.Response to the user. In the event a cache does not exist, we will immediately receive an error. Instead of passing this to the user's error handler, we instead return an error object from subscribe.

We also handle the case where a user deletes the cache mid-subscription. In this case we unsubscribe and pass the error to the error handler. That way a user is not spammed with cache does not exist errors.