chore: remove package files for common, fix imports via node_modules symlink
This commit removes all of the package.json and ts config files
from the common dir since we are not treating it as a package, but
rather just as a directory of files that will be shared via symlink
between the SDKs.

We also need to do a hacky thing for now in the build script, where
we symlink the node_modules dir from the sdk into the common dir
so that imports from the files in common will work.  This won't be
an issue for the packaged artifact since the symlink won't exist.

In a subsequent commit we still need to add a solution for running
the shared tests in the common directory.