feat: publish the node.js client library from the new package structure
This commit updates the publish scripts to publish the node.js
SDK package in addition to the core package.  We add a script
to wait for the core package to be available on node.js, then
update the dependency in the nodejs SDK package to use the newly
released version of core, then publish the nodejs package.