fix: core dependency version for web sdk
The web SDK publish script is modifying the version number for
pre-1.0 releases for now, but prior to this commit it was also
using that same munged version number for specifying it's
dependency on @gomomento/core.  We need to use the real version
number for the core dependency; this commit updates the scripts
to require two version arguments, one for the package version
and one for the core dependency version.