chore: add examples that can be used in docs
We are working on some improvements to our public docs site that
will allow us to pull in code examples from the SDK repos so that
we can still validate the code in CI in the SDK repos.

This commit adds the first few example snippets and ensures that
they get exercised in CI.  In subsequents in the dev docs repo
we will pull this code in and extract the snippets from the
functions for inclusion in the docs.