fix: use a packed version of common integration tests during publish
During publish we were referencing common-integration-tests package
via a relative path.  This meant that it had its own node_modules
directory.  This meant that classes from `sdk-core` were present
in two different node_modules directories, which makes npm very
unhappy when you are doing things like testing than an object is
an instance of a certain class.

In this commit, during the publish phase, we do an npm pack on the
integration tests and then use that as the dep for the SDKs.  That
way they don't try to use anything from the int tests node_modules
dir.