feat: update to more modern version of protogen and grpc-js
This commit updates us to the latest versions of the generated
client-protos, which includes a newer version of protogen and
grpc-js.  This was inspired by this comment related to a bug
that we recently discovered:

https://github.com/momentohq/client-sdk-nodejs/pull/488#issuecomment-1534010405

Haven't run the new test case yet to confirm whether this addresses
it, but upgrading to the latest should be for the best regardless.