feat: introduce API for scoping generated token permissions
This commit tweaks the API for AuthClient.generateAuthToken a
bit further, adding a new 'scope' argument that can be used
to control the permissions of the generated token.

For now there is exactly one pre-built scope that we provide,
currently called AllDataReadWrite.  For console use cases,
it is possible to manually build up a list of restrictions rather
than using this pre-built scope, and thus create a SuperUser token.
End users who are consuming the API will not have permissions to
do this, though, thus we do not currently provide a pre-built
scope object for that.

This commit attempts to shape the API but does not yet wire
up the actual functionality of requesting the token with the
restricted scope; we can go ahead and merge the API changes now
but the backing implementation is blocked until some server-side
changes land.