feat: wire up token permissions and add integration tests
This commit wires up the gRPC plumbing for the generateAuthToken
API to send up the new permissions objects to the server.  Also
adds integration tests to validate which types of tokens are and
are not allowed to call generateAuthToken, depending on what types
of permissions they are trying to assign to the new token.