chore: export ICacheClient from core
Some libraries require unit tests (mocks) with integrations. Therefore
our integrations need to use dependency injection on the interface
`ICacheClient` so we can mock it out in the tests.

Currently we are not exporting `ICacheClient` from the Node.js or web
SDKs. Left alone we would have to separately install @omomento/core to
access ICacheClient, which is less than desirable. After this PR I
will follow up to re-export `ICacheClient` from the SDKs.