feat: create new TopicConfiguration class for configuring TopicClient (#572)
Prior to this commit, the TopicClient was using a CacheClient
Configuration object for configuration. There are many settings
on the CacheClient configuration that aren't relevant for Topics,
and soon the reverse will likely be true as well. This commit changes
the TopicClient constructor to take a new TopicConfiguration object
which will give us the ability to define the configuration API for the
two clients independently going forward.