chore: Create subdirs for web SDK examples
This is similar to what we recently did for the nodejs SDK, where
we created discrete subdirs for the various example projects to
make the dependency management more clear. We are about to start
adding more examples for the web SDK, so this moves the current
ones to a 'cache' subdir.