fix: response type names for ItemGetType and ItemGetTtl, other minor fixes
This commit changes the response type name for ItemGetType and ItemGetTtl
to be prefixed with `Cache` like all of the other response types.

Also changes accessor for ttl to be named `remainingTtlMillis`, to be
consistent with the go SDK and to make it a bit more clear.

Also fixes a few places where we had copy/pasta errors and were returning
ItemGetType responses for ItemGetTtl requests.