fix: Create caches in auth test with session token
In CI the session token does not have the same customer id as the v1 and legacy auth tokens. This means that when testing for example that we can get and set to a cache with a token generated from the session token, we have to create the cache with the session token otherwise it will return Cache Not Found.