feat: add logging implementation of experimental metrics middleware
This commit refactors the "Experimental Metrics CSV" middleware to
be an abstract class that is destination-agnostic. This will allow
us to provide implementations of it that write to other targets
besides CSV. Also included in this PR is an implementation that
writes the metrics to the logger, which is a slightly more scalable
destination for most of our users to consider enabling.