fix: make protobuf types a regular dependency
Prior to this commit, the @types/google-protobuf dep was a dev
dependency rather than a regular dependency.

This meant that when we published the SDK, it wasn't treated as
a required dependency for end users. This is problematic because
our middleware API includes the protobuf `Message` type.

I'm not sure what changed recently; possibly something in the
upstream grpc-js library, but for whatever reason this started
causing build failures in the example projects when trying to
build them against 1.30.0. They were failing with an error
about missing the type definition for `Message`.

This commit makes the types dependency a first-class dep. We
will need to be a little careful about this, it's not ideal to
be exposing types from an upstream dependency in our API signatures
and at some point we might want to create our own wrapper type
that basically redefines the relevant parts of the protobuf
Message type, but this should get the builds working for now.