feat: Add a client factory method that can eagerly connect. (#711)
* feat: Add a client factory method that can eagerly connect.

Extend CacheClientProps to add eagerConnectTimeout, an optional number
that represents the time in seconds to wait for a CacheClient to eagerly
connect to Momento.

Add a factory method that creates a CacheClient. It will eagerly create
connections if eagerConnectTimeout is present, and simply create and
return a client if it is not.

Change the constructors in cache-client.test.ts to use the new create
method. Other tests should be changed in the future.