feat: add a configuration for lambda (#708)
* feat: add a configuration for lambda

Make the number of data clients a cache client creates configurable.
6 is likely too many for lambdas as the extra latency for the first 6
calls isn't worth it in a short-lived environment.

Create a new configuration, InRegion.Lambda, for use inside lambdas.
It only creates 1 data client instead of 6.

* PR fixes

Move the lambda config out of InRegion.

Get rid of the v1 config since it isn't set in stone yet. We will make
one in the future.

Make the numClients unique in the various tests.

Add a test showing numClients can be overridden.