chore: minor re-work to new Middleware decorators to reduce allocations
This commit tweaks the middleware interceptor to only create the
Middleware* decorator objects once, and then re-use them for the
whole chain of middlewares, rather than converting back and forth
between the gRPC type and the decorator type for every middleware
in the chain. This would significantly reduce the number of
objects that needed to be allocated if there were multiple middleware
handlers registered.