feat: introduce `CacheClient.cache` function to curry cache name across API calls (#686)
* feat: introduce `CacheClient.cache` function to curry cache name across API calls

This commit introduces a new function, `CacheClient.cache`. If you call
it, you will get back an instance of a new type, `MomentoCache`. This
object has all of the same APIs that are available on the `CacheClient`,
but it curries the `cacheName` so that you no longer have to specify that
on every API call.