feat: add value accessors on all relevant response classes
This commit includes the following two major changes:
* All relevant `Hit`/`Success` classes now have a `value()` function,
  which returns the data type that we expect to be the most commonly
  used one for the response.
* All relevant `Response` classes now have a `value()` function that
  returns the `Hit`/`Success` subclass's `value()` for the happy path,
  or `undefined` for the `Miss`/`Error` path.

This should simplify a lot of example code, so that we don't have to
call `valueRecord` or similar explicitly specified accessors, and we
can now access `value()` without doing an `instanceof` check first
(even in TypeScript).

NOTE: it is crucially important that we review the choice of data
types for each `value()` function closely and make sure that we
are all aligned on those choices before this is merged.