chore: give vector response types more explicit names, and put them into a namespace
This commit attempts to put a little more intentional organizational
structure in place for the vector response types. This diverges a bit
from the way the other response types are organized, but we should
be able to move the other response types into a structure like this
in the future, while still supporting the old import paths for backward
compat.

I'm not quite satisfied with this yet because I'd like to have a way
to optionally import the types without the `vector.` prefix, so I will
follow up with another commit to address that.