chore: update core exports to simplify access to vector responses
In the previous commit I had grouped the vector response types
into a `vector` namespace, hoping that there would be an easy
way to import the classes directly from within that. However,
I couldn't find one, so this commit also exports them at the
top level of the index so that users can choose whether they
want the `vector.` prefix or not.