chore: disable dependabot and add a script that updates all example deps
This commit disables dependabot for now; it was unsustainable for this
repo with all of the example projects. We can try to enable it again
once they address https://github.com/dependabot/dependabot-core/issues/2178 .

For now we add a script that updates all the example dependency versions.
We can run this by hand after releases, or add a new cron-type Action that
runs it once a day and generates a PR if it produces a diff.