feat: support more vector item metadata value types
In the previous version of the service, we only supported
string-valued metadata. In the latest version, we support the
following value types:

- string
- int
- double
- bool
- list of strings

We update the `VectorItem` and `SearchHit` interfaces to support
metadata of these types (where int and float are both represented as
`number`).

We update `sendUpsertItemBatch` to serialize the metadata
appropriately. We distinguish numbers that are integers vs
floating-point.

Similarly we update `sendSearch` to deserialize the metadata
appropriately.

We add an extra integration test to ensure each of the metadata value
types are stored and retrieved properly.