feat: add mvi score threshold parameter to search method (#1012)
Adds the scoreThreshold parameter to the MVI search method SearchOptions in both Node.js and Web. This feature allows users to filter the search results by distance on the server.

To test we add integration tests that exercise the score threshold on all a dataset and query for each of cosine similarity, inner product, and euclidean similarity.