chore: remove normalization from event loop delay metadata in metrics middleware
I don't think the normalization was doing what we meant for it to do (we were getting
values from users that were showing -19.999 for almost all the percentiles, after we
subtracted the eventLoopDelay of 20). Also, I think it'll be better to have the raw
data from the node engine when we are doing analysis; we can always post-process it.