test: add suffix to identify integration test index names
Because we are seeing indexes leak in integration tests, we
hypothesize a particular integration test or tests may be at fault. To
identify these, we add a suffix to the test name that identifies each
test case.

This is optional and backwards compatible. We may revert this going
forward, though I feel this is a good practice to adopt going forward
across the board.