feat: support configuring multiple grpc connections for leaderboard client
Prior to this commit, the node.js leaderboard client did not honor the
our grpc config setting for "numClients", which is intended to control the
number of TCP connections we make to the server. This commit adds support
for that setting, meaning that users who may exceed 100 concurrent requests
can increase the value in order to get higher throughput.