feat: support array inputs for SortedSetPutElements
Prior to this commit the SortedSetPutElements API would accept a
Map or a Record containing the elements that the caller wished
to add to the sorted set.

A user reported an error when writing JS (not TS) code where they
had attempted to pass in an array of tuples (equivalent to what the
map entries would have looked like) rather than an actual Map.

This commit adds support for passing an array of tuples without
wrapping it in a Map first.