feat: support default configuration for client objects (#1100)
Based on some feedback about example code and related content,
this commit makes the Configuration arguments optional for all
Client constructors, and provides default configurations for
each one. This will allow us to simplify our code examples
so that it is easier to highlight the more interesting and
important parts of the code.