chore: DRY up nodejs grpc channel configuration
Prior to this commit we had several different places in the code where
we were manually constructing an instance of grpc-js's `ChannelOptions`.

This commit creates a helper function to consolidate the construction
in a single spot; we accept an instance of our `GrpcConfiguraion` as input.
As we add support for more of the different grpc settings over time, this
will become increasingly more important.