fix: add support for item sizes up to 5 mb.
Prior to this commit, we were not explicitly setting the max
size for grpc requests and responses. The default was 4mb, which
meant that if a user had requested a limit increase to support
5mb item sizes, the SDK was not actually capable of serving those
items back to them.

This commit adds configuration options to support setting different
values for the max request and response size, and sets the defaults
to 5mb to support those customers.