chore: temporarily allow >= for expected throttling limits (#1165)
We have a test that makes assertions about the cache limits configuration,
both to provide test coverage of the control plane API and to ensure that
we have a sufficient, uniform configuration across the canary environments.

We are currently in the process of raising the limits to allow more parallelization
of tests. This PR temporarily relaxes the constraint for the TPS limit from
equality to `>=` to allow us some flexibility until we know the final value.