feat: add compression for dictionaries (#1223)
Adds compression support for dictionaries, specifically on dictionarySetField, dictionarySetFields, dictionaryGetField, dictionaryGetFields, and dictionaryFetch. We add integration tests on the compression library and refactor the library as well to handle comparing Uint8Arrays better.

In a future PR we ought to refactor the backend to reduce code duplication. With the integration tests in place now, that should be less risky.