feat: remove compression options for dictionary APIs
There is a lot of complexity and sharp edges to enabling compression
on dictionary APIs. The intent for now was only to support compression
for basic scalar APIs. This commit reverts the compression support
for dictionaries (though we retain support for get/set), and we will
add it back in in the future, after more deliberate design, if we
have customers who need it.